move plan state
