+++ src/interface/src/app/resolvers/plan.resolver.spec.ts
@@ -4,7 +4,7 @@ import {
   ParamMap,
   RouterStateSnapshot,
 } from '@angular/router';
-import { PlanState } from '../maplibre-map/plan.state';
+import { PlanState } from '../plan/plan.state';
 import { planResolver } from './plan.resolver';
 import { PlanService } from '@services';
 import { MockProvider } from 'ng-mocks';
